ASP_Assignment_1

+++ Project/Mini_Project/Mini_Project/Admin1/DeleteTrain.cs
@@ -11,24 +11,50 @@ namespace Mini_Project.Admin1
     {
         public static void fn_DeleteTrain()
         {
-            Console.Write("Enter Train Number to cancel: ");
-            int trainNo = int.Parse(Console.ReadLine());
-
-            using (SqlConnection conn = Connection.getconnection())
+            try
             {
-                string deleteClasses = "DELETE FROM TrainClasses WHERE TrainNo = @TrainNo";
-                string deleteTrain = "DELETE FROM Trains WHERE TrainNo = @TrainNo";
+                Console.Write("Enter Train Number to cancel: ");
+                int trainNo = int.Parse(Console.ReadLine());
 
-                SqlCommand cmd1 = new SqlCommand(deleteClasses, conn);
-                SqlCommand cmd2 = new SqlCommand(deleteTrain, conn);
+                using (SqlConnection conn = Connection.getconnection())
+                {
+                    SqlCommand cd = new SqlCommand($"select passengerid from passenger where trainno={trainNo}", conn);
+                    SqlDataReader read = cd.ExecuteReader();
+                    int c = 0;
+                    while (read.Read())
+                    {
+                        c++;
+                    }
+                    read.Close();
+                    if (c == 0)
+                    {
+                        string deleteClasses = "DELETE FROM TrainClasses WHERE TrainNo = @TrainNo";
+                        string deleteTrain = "DELETE FROM Trains WHERE TrainNo = @TrainNo";
 
-                cmd1.Parameters.AddWithValue("@TrainNo", trainNo);
-                cmd2.Parameters.AddWithValue("@TrainNo", trainNo);
+                        SqlCommand cmd1 = new SqlCommand(deleteClasses, conn);
+                        SqlCommand cmd2 = new SqlCommand(deleteTrain, conn);
 
-                cmd1.ExecuteNonQuery();
-                cmd2.ExecuteNonQuery();
+                        cmd1.Parameters.AddWithValue("@TrainNo", trainNo);
+                        cmd2.Parameters.AddWithValue("@TrainNo", trainNo);
 
-                Console.WriteLine("Train and associated classes cancelled successfully.");
+                        cmd1.ExecuteNonQuery();
+                        cmd2.ExecuteNonQuery();
+
+                        Console.WriteLine("Train and associated classes cancelled successfully.");
+                    }
+                    else
+                    {
+                        Console.WriteLine("Cannot delete the train, there are passengers booked for this train");
+                    }
+                }
+            }
+            catch(SqlException ex)
+            {
+                Console.WriteLine("Error: " + ex.Message);
+            }
+            catch(Exception ex)
+            {
+                Console.WriteLine("Error: " + ex.Message);
             }
         }
     }
